Reworked some data types. Removed all the includes of limits.h from sphere code (not zlib and other external stuff). Converted all the SHRT/INT/etc _MIN/MAX defines (declared in limits.h) to the corresponding stdint.h defines. Changed/moved typedefs inside aes.h and other encryption files to not affect sphere files. Moved typedefs from os_windows.h and os_unix.h inside datatypes.h. Fixed wchar for unix. Added signed/unsigned, fixed-size INT8, INT16 and unsigned INT32, INT64 (not to be used regularly, wrote a caveat in datatypes.h).

+++ src/graysvr/CChar.cpp
@@ -1619,7 +1619,7 @@ void CChar::InitPlayer( CClient *pClient, const char *pszCharname, bool bFemale,
 	}
 
 	// Get starting items for the profession / skills.
-	int iProfession = INT_MAX;
+	int iProfession = INT32_MAX;
 	bool bCreateSkillItems = true;
 	switch ( prProf )
 	{
@@ -1666,7 +1666,7 @@ void CChar::InitPlayer( CClient *pClient, const char *pszCharname, bool bFemale,
 	{
 		for ( int i = 1; i < 5; i++ )
 		{
-			int iSkill = INT_MAX;
+			int iSkill = INT32_MAX;
 			switch ( i )
 			{
 				case 1:
@@ -1976,7 +1976,7 @@ do_default:
 						else if( !strnicmp(pszKey, "LAST", 4) )
 						{
 							pszKey += 4;
-							INT64 dwLastTime = INT_MAX, dwCurTime = 0;
+							INT64 dwLastTime = INT32_MAX, dwCurTime = 0;
 
 							for ( size_t iAttacker = 0; iAttacker < m_lastAttackers.size(); ++iAttacker )
 							{
@@ -2358,7 +2358,7 @@ do_default:
 				ptDst.Move( GetDirStr( pszKey ) );
 				CRegionBase * pArea = ptDst.GetRegion( REGION_TYPE_MULTI | REGION_TYPE_AREA );
 				if ( !pArea )
-					sVal.FormatHex( ULONG_MAX );
+					sVal.FormatHex( UINT32_MAX );
 				else
 				{
 					DWORD		wBlockFlags	= 0;
@@ -2727,10 +2727,10 @@ do_default:
 			if ( i >= 0 )
 			{
 				int iVal = s.GetArgVal();
-				if (iVal > SHRT_MAX)
-					iVal = SHRT_MAX;
-				else if (iVal < SHRT_MIN)
-					iVal = SHRT_MIN;
+				if (iVal > INT16_MAX)
+					iVal = INT16_MAX;
+				else if (iVal < INT16_MIN)
+					iVal = INT16_MIN;
 
 				Stat_SetBase(static_cast<STAT_TYPE>(i), static_cast<short>(iVal));
 				return true;
